services/dotlan-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eve-dotlan-service"
version = "1.0.0"
description = "EVE Co-Pilot DOTLAN Scraping Service - System activity, sovereignty, alliance data"
requires-python = ">=3.11"
dependencies = [
    "eve-shared",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
